+++ src/compas_timber/fabrication/step_joint_notch.py
@@ -706,6 +706,28 @@ class StepJointNotch(BTLxProcessing):
 
         return Box(dx, dy, self.mortise_height, box_frame)
 
+    def scale(self, factor):
+        """Scale the parameters of this processing by a given factor.
+
+        Note
+        ----
+        Only distances are scaled, angles remain unchanged.
+
+        Parameters
+        ----------
+        factor : float
+            The scaling factor. A value of 1.0 means no scaling, while a value of 2.0 means doubling the size.
+
+        """
+        self.start_x *= factor
+        self.start_y *= factor
+        self.notch_width *= factor
+        self.step_depth *= factor
+        self.heel_depth *= factor
+        self.strut_height *= factor
+        self.mortise_width *= factor
+        self.mortise_height *= factor
+
 
 class StepJointNotchParams(BTLxProcessingParams):
     """A class to store the parameters of a Step Joint Notch feature.