+++ src/compas_timber/ghpython/components_cpython/CT_Plate/code.py
@@ -60,7 +60,7 @@ class Plate(Grasshopper.Kernel.GH_ScriptInstance):
                 rhino_polyline = rs.coercecurve(geometry)
                 line = curve_to_compas(rhino_polyline)
 
-                plate = CTPlate(line, t, v)
+                plate = CTPlate.from_outline_thickness(line, t, v)
                 plate.attributes["rhino_guid"] = str(guid) if guid else None
                 plate.attributes["category"] = c