+++ CHANGELOG.md
@@ -9,16 +9,59 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added attribute `tolerance` to `TimberModel`.
+* Added `scaled` method to `compas_timber.fabrication.BTLxProcessing` to scale the BTLx parameters.
+* Added `scaled` method to `compas_timber.fabrication.BTLxPart` to scale the BTLx parameters.
+* Added `scale` method to `compas_timber.fabrication.JackRafterCut`.
+* Added `scale` method to `compas_timber.fabrication.Drilling`.
+* Added `scale` method to `compas_timber.fabrication.DoubleCut`.
+* Added `scale` method to `compas_timber.fabrication.Lap`.
+* Added `scale` method to `compas_timber.fabrication.FrenchRidgeLap`.
+* Added `scale` method to `compas_timber.fabrication.Tenon`.
+* Added `scale` method to `compas_timber.fabrication.Mortise`.
+* Added `scale` method to `compas_timber.fabrication.StepJoint`.
+* Added `scale` method to `compas_timber.fabrication.StepJointNotch`.
+* Added `scale` method to `compas_timber.fabrication.DovetailTenon`.
+* Added `scale` method to `compas_timber.fabrication.DovetailMortise`.
+* Added `scale` method to `compas_timber.fabrication.Slot`.
+* Added `scale` method to `compas_timber.fabrication.Pocket`.
+* Added `scale` method to `compas_timber.fabrication.Text`.
 * Added `is_joinery` flag to `BTLxProcessing` to indicate if the processing is a result of joinery operation.
+* Added new `compas_timber.fabrication.LongitudinalCut`.
+* Added tasks `update-gh-header` to update the version in the header of the GH components.
+* Added new `compas_timber.connections.XNotchJoint`.
+* Added a proxy class for `Pocket` BTLx processing for performance optimization. 
+* Added `topology` to class `Joint`.
+* Added `location` to class `Joint`.
+* Added `NBeamKDTreeAnalyzer` to `compas_timber.connections`.
+* Added `TripletAnalyzer` to `compas_timber.connections`.
+* Added `QuadAnalyzer` to `compas_timber.connections`.
+* Added `CompositeAnalyzer` to `compas_timber.connections`.
+* Added method `connect_adjacent_beams` to `TimberModel`.
+* Added `PlateJoint`.
+* Added `PlateButtJoint`.
+* Added `PlateMiterJoint`.
+* Added `PlateConnectionSolver`.
+* Added generic `ButtJoint` class from which `TButtJoint` and `LButtJoint` inherit.
 
 ### Changed
 
+* BTLx Write now considers the `TimberModel.tolerance` attribute and scales parts and processings it when units are set to meters.
+* Added missing `__data__` to `compas_timber.fabrication.Drilling`.
+* Added missing `__data__` to `compas_timber.fabrication.Slot`.
 * Fixed `TypeError` when deepcopying beams with `debug_info` on them.
 * Processings which are not the result of joinery are now serialized with `TimberElement`.
 * Fixed visualization bug in `Plate` due to loft resulting in flipped volume.
 * Fixed a few bugs in the `WallPopulator` workflow including GH component updates.
+* Renamed `NullJoint` to `GenericJoint`.
+* Fixed bug in show_ref_faces GH component.
+* `BTLxProcessing.ref_side_index` defaults to `0` if not set, instead of the invalid `None`.
 
 ### Removed
+
+* Removed Grasshopper after-install plugin. Components should be installed via Rhino's Plugin Manager.
+* Removed `get_face_most_towards_beam` from `Joint` as not used anywhere.
+* Removed `get_face_most_ortho_to_beam` from `Joint` as not used anywhere.
 
 
 ## [0.16.2] 2025-05-07
@@ -30,6 +73,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed max recursion depth error when copying `TimberModel`/`Beam` with proxy processings.
 
 ### Removed
+
 
 
 ## [0.16.1] 2025-04-30